% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_article.R
\name{acm}
\alias{acm}
\title{Association for Computing Machinery (ACM) format.}
\usage{
acm()
}
\description{
Format for creating an Association for Computing Machinery (ACM) articles.
Adapted from
\href{http://www.acm.org/publications/article-templates/proceedings-template.html}{http://www.acm.org/publications/article-templates/proceedings-template.html}.
}

