% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_chi_ea_article.R
\name{acm_chi_ea_article}
\alias{acm_chi_ea_article}
\title{ACM CHI Extended Abstracts (Conference on Human Factors in Computing Systems) format.
This version uses v. 1.55 of the ACM Master LaTeX template.}
\usage{
acm_chi_ea_article(..., keep_tex = TRUE, citation_package = "natbib")
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
CHI has two main submission formats, 'proceedings' and 'extended abstracts';
This template is for CHI Extended Abstracts.
Format strictly follows the ACM Master LaTeX template.
ACM's notes here: https://www.acm.org/publications/proceedings-template
Devevelopment version of the ACM template: https://github.com/borisveytsman/acmart
}
\examples{
\dontrun{
rmarkdown::draft("MyArticle.Rmd", template = "asa_article", package = "rticles")
}

}
