% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_article.R
\name{acs}
\alias{acs}
\title{Journal of the American Chemical Society (ACS) format.}
\usage{
acs(keep_tex = TRUE)
}
\arguments{
\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\description{
Format for creating an American Chemical Society (ACS) articles. Adapted from \href{http://pubs.acs.org/page/4authors/submission/tex.html}{http://pubs.acs.org/page/4authors/submission/tex.html}.
}

