% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aea_article.R
\name{aea_article}
\alias{aea_article}
\title{American Economic Association journal submissions.}
\usage{
aea_article(..., keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"))
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to the American Economic Association (AER, AEJ, JEL, PP).
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "aea_article", package = "rticles")
}

}
