% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agu_article.R
\name{agu_article}
\alias{agu_article}
\title{American Geophysical Union (AGU) format}
\usage{
agu_article(..., keep_tex = TRUE, citation_package = "natbib",
  highlight = NULL, md_extensions = c("-autolink_bare_uris",
  "-auto_identifiers"))
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document()}.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}
}
\value{
An R Markdown output format.
}
\description{
Format for creating a American Geophysical Union (AGU) article. Adapted from
\url{https://publications.agu.org/author-resource-center/checklists-and-templates/}.
}
\examples{
\dontrun{
rmarkdown::draft("MyArticle.Rmd", template = "agu_article", package = "rticles")
}
}
