% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amq_article.R
\name{amq_article}
\alias{amq_article}
\title{Format pour Bulletin de l'AMQ.}
\usage{
amq_article(..., latex_engine = "xelatex", keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"), fig_caption = TRUE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", and "xelatex".}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{fig_caption}{\code{TRUE} to render figures with captions}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Ce format a été adapté du format du bulletin de l'AMQ
}
\examples{
\dontrun{
rmarkdown::draft("MyArticle.Rmd", template = "amq_article", package = "rticles")
}

}
