% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aps_article.R
\name{aps_article}
\alias{aps_article}
\title{American Physical Society (APS) Journal format.}
\usage{
aps_article(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris"),
  fig_caption = TRUE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{fig_caption}{\code{TRUE} to render figures with captions}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating an American Physical Society (APS) Journal articles.
Adapted from
\href{https://journals.aps.org/revtex/revtex-faq}{https://journals.aps.org/revtex/revtex-faq}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "aps_article", package = "rticles")
}

}
