% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asa_article.R
\name{asa_article}
\alias{asa_article}
\title{American Statistical Association (ASA) Journal format.}
\usage{
asa_article(..., keep_tex = TRUE, citation_package = "natbib")
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}
}
\description{
This format was adapted from The American Statistican format, but it
should be fairly consistent across ASA journals
}
