% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinformatics_article.R
\name{bioinformatics_article}
\alias{bioinformatics_article}
\title{Bioinformatics journal format.}
\usage{
bioinformatics_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to Bioinformatics journals.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "bioinformatics_article", package = "rticles")
}

}

