% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmc_article.R
\name{bmc_article}
\alias{bmc_article}
\title{BioMed Central journal format.}
\usage{
bmc_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to bmc journals.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "bmc_article", package = "rticles")
}

}

