% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conll_article.R
\name{conll_article}
\alias{conll_article}
\title{CoNLL paper format}
\usage{
conll_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
CoNLL paper format
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "conll_article", package = "rticles")
}

}
