% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_article.R
\name{copernicus_journal_abbreviations}
\alias{copernicus_journal_abbreviations}
\title{Copernicus journal abbreviations}
\usage{
copernicus_journal_abbreviations(name = "*")
}
\arguments{
\item{name}{A regular expression to filter the by the journal name, see \code{pattern} in \code{\link[base]{grep}}; defaults to \code{*}.}
}
\description{
Get the journal abbreviation for all journals supported by the copernicus article template.
}
\examples{
names(copernicus_journal_abbreviations())
copernicus_journal_abbreviations(name = "Science Data")

}
