% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctex.R
\name{ctex}
\alias{ctex}
\title{A PDF format for documents based on the LaTeX package \pkg{ctex}}
\usage{
ctex(..., latex_engine = "xelatex")
}
\arguments{
\item{..., latex_engine}{Passed to
\code{markdown::\link[rmarkdown]{pdf_document}()}.}
}
\description{
\code{ctex()} is a wrapper function for \code{rmarkdown::pdf_document()} and
changed the default values of two arguments \code{template} and
\code{latex_engine} so it works better with the \pkg{ctex} package.
}
