% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ctex.R
\name{ctex_template}
\alias{ctex_template}
\title{A LaTeX template for Chinese documents}
\usage{
ctex_template()
}
\description{
This function returns the path to a LaTeX template for Chinese documents
using the ctex package.
}

