% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degruyter_article.r
\name{degruyter_article}
\alias{degruyter_article}
\title{De Gruyter Journal format.}
\usage{
degruyter_article(..., keep_tex = TRUE, citation_package = "natbib")
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
This format was adapted from JQAS format, but it
should be fairly consistent across DeGruyter journals
}
\details{
This package contains a template for writing articles
in R Markdown for De Gruyter journals such as Journal of
Quantitative Analysis in Sports (JQAS).

Files from the "De Gruyter LaTeX Template Download" .zip file
from \url{https://www.degruyter.com/view/j/jqas}
were used for creating this template.  In particular,

\enumerate{
  \item \code{skeleton/dgjournal.sty} and \code{skeleton/DeGruyter.bst} from that
 \code{.zip} file are used for styling
  \item \code{skeleton/preambleDeGruyter.tex} contains a preamble from
 the example document \code{dg_template.tex} in that .zip file.
 \item The \code{skeleton.rmd} file contains example content copied from \code{dg_template.tex} as well.
}

Note that for De Gruyter journal submissions,
front matter can not be included.
Because of this, \code{title:}, \code{author:},
and other similar lines are commented out in \code{skeleton.rmd}.

In an Arxiv submission, this front matter may be desired.
You can choose to knit with the output format \code{pdf_document}
instead of \code{pdf_document2},
which will not use the \code{dgjournal.sty} styling,
and will allow front matter.  The lines with the front matter
can be uncommented and they will appear in the .pdf output.
Note that the \code{pdf_document} output option has its own
preamble file \code{preambleArxiv.tex} which can be modified as desired.
}
\examples{
\dontrun{
rmarkdown::draft("MyArticle.Rmd", template = "degruyter_article", package = "rticles")
}

}
