% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drreport.R
\name{drdocumentsetup}
\alias{drdocumentsetup}
\title{Setup an rmarkdown document for Digital Reasoning defaults}
\usage{
drdocumentsetup()
}
\description{
This function should be called in the first, setup, chunk of the document.
}

