% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_ebook.R, R/utils.R
\name{drreport}
\alias{drreport}
\alias{margin_note}
\alias{newthought}
\alias{quote_footer}
\alias{sans_serif}
\title{Digital Reasoning Full-Length Report}
\usage{
drreport(fig_width = 4, fig_height = 2.5, fig_crop = TRUE, dev = "pdf",
  highlight = "default", ...)

newthought(text)

margin_note(text, icon = "&#8853;")

quote_footer(text)

sans_serif(text)
}
\arguments{
\item{text}{A character string to be presented as a \dQuote{new thought}
(using small caps), or a margin note, or a footer of a quote}

\item{icon}{A character string to indicate there is a hidden margin note when
the page width is too narrow (by default it is a circled plus sign)}
}
\description{
In the style of Edward Tufte, with table of contents, title page, etc.
}
\details{
\code{newthought()} can be used in inline R expressions in R
  Markdown (e.g. \samp{`r newthought(Some text)`}), and it works for both
  HTML (\samp{<span class="newthought">text</span>}) and PDF
  (\samp{\\newthought{text}}) output.

\code{margin_note()} can be used in inline R expressions to write a
  margin note (like a sidenote but not numbered).

\code{quote_footer()} formats text as the footer of a quote. It puts
  \code{text} in \samp{<footer></footer>} for HTML output, and
  after \samp{\\hfill} for LaTeX output (to right-align text).

\code{sans_serif()} applies sans-serif fonts to \code{text}.
}
\examples{
newthought('In this section')
}

