% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsevier_article.R
\name{elsevier}
\alias{elsevier}
\title{Elsevier journal format.}
\usage{
elsevier(...)
}
\arguments{
\item{...}{additional arguments to \code{rmarkdown::pdf_document}}
}
\description{
Format for creating submissions to Elsevier journals. Adapted from \href{https://www.elsevier.com/authors/author-schemas/latex-instructions}{https://www.elsevier.com/authors/author-schemas/latex-instructions}.
}

