% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erc_article.R
\name{erc_article}
\alias{erc_article}
\title{ERC format}
\usage{
erc_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for EU funding applications.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "erc_article", package = "rticles")
}

}
