% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_booklet_de.R
\name{et_booklet_de}
\alias{et_booklet_de}
\title{E.Tutorial Booklet for online distribution}
\usage{
et_booklet_de(..., keep_tex = F)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
E.Tutorial Booklet for online distribution
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "et_booklet_de", package = "eTutorialsLayout")
}

}

