% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_article.R
\name{frontiers_article}
\alias{frontiers_article}
\title{Frontiers open access journal format.}
\usage{
frontiers_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\description{
Format for creating Frontiers journal articles. Adapted from
\url{http://home.frontiersin.org/about/author-guidelines}.
}
