% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigascience_article.R
\name{gigascience_article}
\alias{gigascience_article}
\title{GigaScience journal format.}
\usage{
gigascience_article(
  ...,
  keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris")
)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}
}
\value{
R Markdown output format to pass to \code{\link[=rmarkdown:render]{rmarkdown:render()}}
}
\description{
Format for creating submissions to the GigaScience journal. Adapted from GigaScience Overleaf template
\url{https://www.overleaf.com/latex/templates/template-for-gigascience-journal-manuscript-submissions/shgtrssvbjhs}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "gigascience_article", package = "rticles")
}

}
