% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieee_article.R
\name{ieee_article}
\alias{ieee_article}
\title{IEEE Transactions journal format.}
\usage{
ieee_article(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris"))
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to IEEE Transaction journals. Adapted from
\href{http://www.ieee.org/publications_standards/publications/authors/author_templates.html}{http://www.ieee.org/publications_standards/publications/authors/author_templates.html}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "ieee_article", package = "rticles")
}

}

