% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacsm_article.R
\name{jacsm_article}
\alias{jacsm_article}
\title{Journal of Applied Computer Science Methods (JACSM) article format.}
\usage{
jacsm_article(
  ...,
  keep_tex = TRUE,
  chunk_highlight = FALSE,
  chunk_prompt = FALSE,
  chunk_comment = NA
)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF.
Note that this argument does not control whether to keep the auxiliary
files (e.g., \file{.aux}) generated by LaTeX when compiling \file{.tex} to
\file{.pdf}. To keep these files, you may set \code{options(tinytex.clean =
FALSE)}.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating JACSM articles. Based on
\href{https://www.degruyter.com/view/j/jacsm}{https://www.degruyter.com/view/j/jacsm}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "jacsm_article", package = "rticles")
}

}
