% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jds_article.R
\name{jds_article}
\alias{jds_article}
\title{Journal of Data Science (JDS)}
\usage{
jds_article(
  review = FALSE,
  keep_tex = TRUE,
  citation_package = "natbib",
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{review}{Use options for document classe associated to review status.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{default} if neither package is to be used,
which means citations will be processed via the command
\command{pandoc-citeproc}.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{Arguments passed to \code{pdf_document()}.}
}
\description{
Format for creating Journal of Data Science (JDS) articles. Adapted from
\url{https://jds-online.org/journal/JDS/information/submit-your-article}.
}
