% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{journals}
\alias{journals}
\title{List available journals}
\usage{
journals()
}
\value{
a vector with the journal names of format available in rticles
}
\description{
The function \code{\link[=journals]{journals()}} will list available journal names in the \strong{rticles}
package. These names can be useful in two ways :
\itemize{
\item You can add \verb{_article} suffix to get the format (i.e \code{\link[=rjournal_article]{rjournal_article()}}).
Only exception is for the \code{\link[=ctex]{ctex()}} format.
\item You can use the name directly in the \code{template} argument of
\code{\link[rmarkdown:draft]{rmarkdown::draft()}}.
}
}
\examples{
rticles::journals()
}
