% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jss_article.R
\name{jss_article}
\alias{jss_article}
\title{Journal of Statistical Software (JSS) format.}
\usage{
jss_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF.
Note that this argument does not control whether to keep the auxiliary
files (e.g., \file{.aux}) generated by LaTeX when compiling \file{.tex} to
\file{.pdf}. To keep these files, you may set \code{options(tinytex.clean =
FALSE)}.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating a Journal of Statistical Software (JSS) articles. Adapted
from
\href{http://www.jstatsoft.org/about/submissions}{http://www.jstatsoft.org/about/submissions}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "jss_article", package = "rticles")
}

}
