% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lncs_article.R
\name{lncs_article}
\alias{lncs_article}
\title{Lecture Notes in Computer Science (LNCS) proceedings format.}
\usage{
lncs_article(..., keep_tex = TRUE, chunk_highlight = FALSE,
  chunk_prompt = FALSE, chunk_comment = NA)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating Lecture Notes in Computer Science (LNCS) proceedings articles. Adapted
from
\href{http://www.springer.com/comp/lncs/authors.html}{http://www.springer.com/comp/lncs/authors.html}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "lncs_article", package = "rticles")
}

}

