% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpi_article.R
\name{mdpi}
\alias{mdpi}
\alias{mdpi_article}
\alias{mdpi_journals}
\title{MDPI Journal}
\usage{
mdpi_article(
  ...,
  keep_tex = TRUE,
  latex_engine = "pdflatex",
  pandoc_args = NULL,
  citation_package = "natbib"
)

mdpi_journals()
}
\description{
Format for creating submissions to Multidisciplinary Digital Publishing Institute (MDPI) journals.
Adapted from \url{https://www.mdpi.com/authors/latex}.
}
\details{
Possible arguments for the YAML header are:
\itemize{
\item \code{title} title of the manuscript
\item \code{author} list of authors, containing \code{name}, \code{affil}, and \code{orcid} (optional)
\item \code{affiliation} list containing \code{num}, \code{address}, and \code{email} for defining \code{author} affiliations
\item \code{authorcitation} string with last name and first intial of authors as expected to be shown in a reference
\item \code{firstnote} can include \code{firstnote} through \code{eightnote} that correspond to footnote marks in \code{affil}
\item \code{correspondence} contact information of the corresponding author
\item \code{journal} short name (case sensitive) of the journal (see list in template or with \code{mdpi_journals()}).
\item \code{type} usually "article" but see template for options
\item \code{status} usually "submit"
\item \code{simplesummary} optional, may depend on specific journal
\item \code{abstract} abstract, limited to 200 words
\item \code{keywords} 3 to 10 keywords seperated with a semicolon
\item \code{acknowledgement} acknowledgement backmatter (optional)
\item \code{authorcontributions} report authorship contributions (optional)
\item \code{funding} research funding statement
\item \code{institutionalreview} IRB statements (optional)
\item \code{informedconsent} Informed consent statements for human research (optional)
\item \code{dataavailability} Links to datasets or archives (optional)
\item \code{conflictsofinterest} Conflict of interest statement (see journal  instructions)
\item \code{sampleavailability} Sample availability statement (optional)
\item \code{supplementary} Supplementary data statement, see template for example (optional)
\item \code{abbreviations} list of abbreviations containing \code{short} and \code{long}
\item \code{bibliography} BibTeX \code{.bib} file
\item \code{appendix} name of appendix tex file
\item \code{endnote} boolean, if \code{TRUE} will print list of endnotes if included in text (optional)
\item \code{header-includes}: custom additions to the header, before the \verb{\\begin\{document\}} statement
\item \code{include-after}: for including additional LaTeX code before the \verb{\\end\{document\}} statement
}
}
