% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih_r01_grant.R
\name{nih_r01_grant}
\alias{nih_r01_grant}
\title{NIH R01 format.}
\usage{
nih_r01_grant(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating R01 submissions to the NIH.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "nih_r01_grant", package = "rticles")
}

}

