% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opmi_article.R
\name{opmi_article}
\alias{opmi_article}
\title{OpenMind article format}
\usage{
opmi_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
OpenMind article format
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "opmi_article", package = "rticles")
}

}
