% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oup_article.R
\name{oup_article}
\alias{oup_article}
\title{Oxford University Press.}
\usage{
oup_article(
  ...,
  keep_tex = TRUE,
  oup_version = 0,
  md_extensions = c("-autolink_bare_uris"),
  journal = NULL,
  pandoc_args = NULL,
  number_sections = TRUE,
  number_lines = FALSE,
  number_lines_options = NULL,
  citation_package = ifelse(oup_version == 0, "default", "natbib"),
  papersize = c("large", "medium", "small"),
  document_style = c("contemporary", "modern", "traditional"),
  namedate = FALSE,
  onecolumn = FALSE
)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{oup_version}{For \code{oup_article}, set to 0 to use the 2009 OUP CLS style for document formatting or set to 1 to use the newer 2020 OUP CLS style package available on CTAN. Defaults to 0.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{number_sections}{\code{TRUE} to number section headings}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{default} if neither package is to be used,
which means citations will be processed via the command
\command{pandoc-citeproc}.}

\item{papersize}{For \code{oup_article}, \code{oup_version=1}, one of "large", "medium", or "small" setting output page size. Defaults to "large".}

\item{document_style}{For \code{oup_article}, \code{oup_version=1}, one of "contemporary", "modern", or "traditional" setting overall style of document. Defaults to "contemporary".}

\item{namedate}{For \code{oup_article}, \code{oup_version=1}, a logical variable indicating if natbib citations should be in name-date format. Defaults to \code{FALSE}.}

\item{onecolumn}{For \code{oup_article}, \code{oup_version=1}, a logical variable indicating if one column formatting should be used. Defaults to \code{FALSE}.}
}
\description{
Format for creating submissions to many Oxford University Press journals.
Adapted from
\url{https://academic.oup.com/journals/pages/authors/preparing_your_manuscript}
and \url{https://academic.oup.com/icesjms/pages/General_Instructions}. and the
\code{oup-authoring-template} available on CTAN at
\url{https://www.ctan.org/pkg/oup-authoring-template}.
}
\details{
Note that for
\itemize{
\item \code{oup_version=0}, \code{citation_package="default"} by default,
\item \code{oup_version=1}, \code{citation_package="natbib"} by default and
\code{citation_package="biblatex"} is not supported.
}
}
