% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plos_article.R
\name{plos_article}
\alias{plos_article}
\title{Public Library of Science (PLOS) format.}
\usage{
plos_article(keep_tex = TRUE, includes = NULL)
}
\arguments{
\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}
}
\description{
Format for creating Public Library of Science (PLOS) articles. Adapted from \href{http://journals.plos.org/plosone/s/latex}{http://journals.plos.org/plosone/s/latex}.
}

