% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjournal_article.R
\name{rjournal_article}
\alias{rjournal_article}
\title{R Journal format.}
\usage{
rjournal_article()
}
\description{
Format for creating R Journal articles. Adapted from \href{https://journal.r-project.org/submissions.html}{https://journal.r-project.org/submissions.html}.
}

