% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsos_article.R
\name{rsos_article}
\alias{rsos_article}
\title{Royal Society Open Science journal format.}
\usage{
rsos_article(
  ...,
  keep_tex = TRUE,
  latex_engine = "xelatex",
  pandoc_args = NULL,
  includes = NULL,
  fig_crop = TRUE
)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", "xelatex" and "tectonic".}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{fig_crop}{Whether to crop PDF figures with the command
\command{pdfcrop}. This requires the tools \command{pdfcrop} and
\command{ghostscript} to be installed. By default, \code{fig_crop = TRUE}
if these two tools are available.}
}
\description{
Format for creating submissions to Royal Society Open Science journals.
}
\author{
Thierry Onkelinx, \email{thierry.onkelinx@inbo.be}
}
