% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/springer_article.R
\name{springer_article}
\alias{springer_article}
\title{Springer Journal format.}
\usage{
springer_article(..., keep_tex = TRUE, citation_package = "none")
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
This format was adapted from the Springer Macro package for Springer
Journals.
}
\examples{
\dontrun{
rmarkdown::draft("MyArticle.Rmd", template = "springer_article", package = "rticles")
}

}
