% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjournal_article.R
\name{template_pandoc}
\alias{template_pandoc}
\title{Render a pandoc template.}
\usage{
template_pandoc(metadata, template, output, verbose = FALSE)
}
\arguments{
\item{metadata}{A named list containing metadata to pass to template.}

\item{template}{Path to a pandoc template.}

\item{output}{Path to save output.}
}
\value{
(Invisibly) The path of the generate file.
}
\description{
This is a hacky way to access the pandoc templating engine.
}
\examples{
x <- rticles:::template_pandoc(
  list(preamble = "\%abc", filename = "wickham"),
  rticles:::find_resource("rjournal_article", "RJwrapper.tex"),
  tempfile()
)
if (interactive()) file.show(x)
}

