% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_article.R
\name{tf_article}
\alias{tf_article}
\title{Taylor and Fracis journal format.}
\usage{
tf_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to a Taylor & Francis journal. Adapted from
\href{http://www.tandf.co.uk/journals/authors/InteractCADLaTeX.zip}
{http://www.tandf.co.uk/journals/authors/InteractCADLaTeX.zip}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "tf_article", package = "rticles")
}

}
