% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_ebook.R
\name{tufte_ebook}
\alias{tufte_ebook}
\title{Tufte e-Book format (PDF)}
\usage{
tufte_ebook(toc = TRUE, toc_depth = 3, number_sections = TRUE,
  fig_width = 4, fig_height = 2.5, fig_crop = TRUE,
  highlight = "default", keep_tex = FALSE, latex_engine = "xelatex",
  includes = NULL, pandoc_args = NULL, ...)
}
\arguments{
\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", and "xelatex".  Note that lualatex may have
problems with text spacing, and that pdflatex may have memory issues
when using tikzDevice.}
}
\description{
Template for creating an e-book based on the style of
Edward R. Tufte and Richard Feynman.
}

