% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tuftish_ebook.R
\name{tuftish_ebook}
\alias{tuftish_ebook}
\title{Tuftish e-Book format (PDF)}
\usage{
tuftish_ebook(fig_width = 4, fig_height = 2.5, fig_crop = TRUE,
  highlight = "default", keep_tex = FALSE, includes = NULL,
  pandoc_args = NULL, toc = TRUE, toc_depth = 3,
  latex_engine = "lualatex", number_sections = TRUE, ...)
}
\description{
Template for creating an e-book based on the style of
Edward R. Tufte and Richard Feynman.
}

