% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_abstract.R
\name{use_r_abstract}
\alias{use_r_abstract}
\title{useR conference abstract format.}
\usage{
use_r_abstract()
}
\description{
Format for useR conference abstracts. Adapted from \href{http://user2016.org/}{http://user2016.org/}.
}

