% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsc_article.R
\name{wsc_article}
\alias{wsc_article}
\title{wsc format.}
\usage{
wsc_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating a wsc
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "jss_article", package = "rticles")
}

}

