% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_article.R
\name{acm_article}
\alias{acm_article}
\title{Association for Computing Machinery (ACM) format.}
\usage{
acm_article(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\description{
Format for creating an Association for Computing Machinery (ACM) articles.
Adapted from
\href{http://www.acm.org/publications/article-templates/proceedings-template.html}{http://www.acm.org/publications/article-templates/proceedings-template.html}.
}

