% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_article.R
\name{acs_article}
\alias{acs_article}
\title{Journal of the American Chemical Society (ACS) format.}
\usage{
acs_article(..., fig_caption = TRUE, keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating an American Chemical Society (ACS) articles. Adapted from \href{http://pubs.acs.org/page/4authors/submission/tex.html}{http://pubs.acs.org/page/4authors/submission/tex.html}.
}

