% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agu_article.R
\name{agu_article}
\alias{agu_article}
\title{American Geophysical Union (AGU) format}
\usage{
agu_article(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating a American Geophysical Union (AGU) article.
Adapted from
\href{https://publications.agu.org/author-resource-center/checklists-and-templates/}{https://publications.agu.org/author-resource-center/checklists-and-templates/}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "agu_article", package = "rticles")
}

}
