% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams_article.R
\name{ams_article}
\alias{ams_article}
\title{American Meteorological Society journals format.}
\usage{
ams_article(
  ...,
  keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris", "-auto_identifiers")
)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}. \strong{Note}: \code{extra_dependencies} are not
allowed as Copernicus does not support additional packages included via \code{\\usepackage{}}.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF.
Note that this argument does not control whether to keep the auxiliary
files (e.g., \file{.aux}) generated by LaTeX when compiling \file{.tex} to
\file{.pdf}. To keep these files, you may set \code{options(tinytex.clean =
FALSE)}.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition of R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}
}
\value{
An R Markdown output format.
}
\description{
Format for creating submissions to American Meteorological Society journals.
}
\details{
This was adapted from
\url{https://www.ametsoc.org/index.cfm/ams/publications/author-information/latex-author-info/}.

The template require some default knitr option to be change:
\itemize{
\item \code{echo = FALSE} as no R code should be shown
\item \code{fig.path = ""} as all directory paths need to be removed from figure names
\item \code{out.extra = ""} to force figure labels with knitr
}
}
\examples{
\dontrun{
library("rmarkdown")
draft("MyArticle.Rmd", template = "ams", package = "rticles")
render("MyArticle/MyArticle.Rmd")
}
}
