% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asa_article.R
\name{asa_article}
\alias{asa_article}
\title{American Statistical Association (ASA) Journal format.}
\usage{
asa_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
This format was adapted from The American Statistican format, but it
should be fairly consistent across ASA journals
}
\examples{
\dontrun{
rmarkdown::draft("MyArticle.Rmd", template = "asa_article", package = "rticles")
}

}
