% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_article.R
\name{copernicus_article}
\alias{copernicus_article}
\alias{copernicus_journal_abbreviations}
\title{Copernicus journals format.}
\usage{
copernicus_article(..., keep_tex = TRUE, citation_package = "natbib",
  base_format = rmarkdown::pdf_document,
  md_extensions = c("-autolink_bare_uris", "-auto_identifiers"))

copernicus_journal_abbreviations(journal_name = "*")
}
\arguments{
\item{...}{Additional arguments to \code{base_format}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}

\item{base_format}{The function to use for the base format of the article.
By default, this is \code{rmarkdown::pdf_document}, but to use bookdown's
cross-referencing feature, this can be set to \code{bookdown::pdf_document2}}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{journal_name}{A regular expression to filter the by the journal name, see \code{pattern} in \code{\link[base]{grep}}; defaults to \code{*}.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to Copernicus journals.
}
\details{
This was adapted from
\href{https://publications.copernicus.org/for_authors/manuscript_preparation.html}{https://publications.copernicus.org/for_authors/manuscript_preparation.html}.

An number of required and optional manuscript sections, e.g. \code{acknowledgements}, \code{competinginterests}, or \code{authorcontribution}, must be declared using the respective properties of the R Markdown header - see skeleton file.

\strong{Version:} Based on copernicus_package.zip in the version 5.0, 21 March 2018, using \code{copernicus.cls} in version 8.67, 30 January 2018

\strong{Copernicus journal abbreviations:} You can use the function \code{copernicus_journal_abbreviations()} to get the journal abbreviation for all journals supported by the copernicus article template.

\strong{Important note:} The online guidelines by Copernicus are the official resource.
Copernicus is not responsible for the community contributions made to support the template in this package.
Copenicus converts all typeset TeX files into XML, the expressions and markups have to be highly standardized.
Therefore, please keep the following in mind:

\itemize{
  \item Please provide only one figure file for figures with several panels, and please do not use \code{\\subfloat} or similar commands.
  \item Please use only commands in which words, numbers, etc. are within braces (e.g. \code{\\textrm{TEXT}} instead of \code{{\\rm TEXT}}).
  \item For algorithms, please use the syntax given in template.tex or provide your algorithm as a figure.
  \item Please do not define new commands.
  \item The most commonly used packages (\code{\\usepackage{}}) are integrated in the copernicus.cls. Some other packages often used by the community are defined in template.tex. Please do not insert additional ones in your *.tex file.
  \item Spaces in labels (\code{\\label{}}) are not allowed; please make sure that no label name is assigned more than once.
  \item Please do not use \code{\\paragraph{}}; only \code{\\subsubsection{}} is allowed.
  \item It is not possible to add tables in colour.
}
}
\note{
If you use \code{rmarkdown::pdf_document()}, all internal references (i.e. tables and figures) must use \code{\\ref\{\}} whereas with \code{bookdown::pdf_document2()}, you can additionally use \code{\\@ref()}.
}
\examples{
names(copernicus_journal_abbreviations())
copernicus_journal_abbreviations(journal_name = "Science Data")
\dontrun{
library("rmarkdown")
draft("MyArticle.Rmd", template = "copernicus_article", package = "rticles")
render("MyArticle/MyArticle.Rmd")
}

}
\references{
Manuscript preparation guidelines for authors.
\url{https://publications.copernicus.org/for_authors/manuscript_preparation.html}
}
