% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsevier_article.R
\name{elsevier}
\alias{elsevier}
\title{elsevier
use the Elsevier template}
\usage{
elsevier(...)
}
\arguments{
\item{...}{additional arguments to \code{rmarkdown::pdf_document}}
}
\description{
elsevier
use the Elsevier template
}

