% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_article.R
\name{frontiers_article}
\alias{frontiers_article}
\title{Frontiers open access journal format.}
\usage{
frontiers_article(keep_tex = TRUE, includes = NULL)
}
\arguments{
\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}
}
\description{
Format for creating Frontiers journal articles. Adapted from \href{http://home.frontiersin.org/about/author-guidelines}{http://home.frontiersin.org/about/author-guidelines}.
}

