% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_article.R
\name{frontiers_article}
\alias{frontiers_article}
\title{Frontiers open access journal format.}
\usage{
frontiers_article(..., highlight = "tango", keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating Frontiers journal articles. Adapted from
\href{http://home.frontiersin.org/about/author-guidelines}{http://home.frontiersin.org/about/author-guidelines}.
}

