% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joss_article.R
\name{joss_article}
\alias{joss_article}
\title{Journal of Open Source Software (JOSS) format.}
\usage{
joss_article(journal = "JOSS", keep_md = TRUE, latex_engine = "xelatex", ...)
}
\arguments{
\item{journal}{one of "JOSS" or"JOSE"}

\item{keep_md}{Whether to retain the intermediate markdown and images.
Defaults to TRUE.}

\item{latex_engine, ...}{Arguments passed to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}}
}
\description{
Format for creating a Journal of Open Source Software (JOSS) or Journal of
Open Source Education (JOSE) articles. Adapted
from \url{https://github.com/openjournals/whedon}. As these journals take
articles as markdown, this format can be used to generate markdown from
R Markdown and to locally preview how the article will appear as PDF.
}
\details{
The following variables may be set in YAML metadata to populate fields in the
article PDF, but are only necessary fo local preview:
\verb{formatted_doi, citation_author, year, volume, issue, page, submitted, published, review_url, repository}, and \code{archive_doi}.
}
