% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jss_article.R
\name{jss_article}
\alias{jss_article}
\title{Journal of Statistical Software (JSS) format.}
\usage{
jss_article()
}
\description{
Format for creating a Journal of Statistical Software (JSS) articles. Adapted from \href{http://www.jstatsoft.org/about/submissions}{http://www.jstatsoft.org/about/submissions}.
}

