% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jss_article.R
\name{jss_article}
\alias{jss_article}
\title{Journal of Statistical Software (JSS) format.}
\usage{
jss_article(
  ...,
  keep_tex = TRUE,
  citation_package = "natbib",
  pandoc_args = NULL
)
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{default} if neither package is to be used,
which means citations will be processed via the command
\command{pandoc-citeproc}.}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\description{
Format for creating a Journal of Statistical Software (JSS) articles. Adapted
from \url{https://www.jstatsoft.org/about/submissions}.
}
