% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpi_article.R
\name{mdpi_article}
\alias{mdpi_article}
\title{Multidisciplinary Digital Publishing Institute (MDPI) Journal format.}
\usage{
mdpi_article(..., keep_tex = TRUE, citation_package = "natbib")
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
This format was adapted from Entropy format, but it
should be fairly consistent across ASA journals
}
\examples{
\dontrun{
rmarkdown::draft("MyArticle.Rmd", template = "mdpi_article", package = "rticles")
}

}
