% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpi_article.R
\name{mdpi_article}
\alias{mdpi_article}
\title{MDPI journal format.}
\usage{
mdpi_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\description{
Format for creating submissions to Multidisciplinary Digital Publishing
Institute (MDPI) journals. Adapted from
\url{http://www.mdpi.com/authors/latex}.
}
