% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpi_article.R
\name{mdpi_article}
\alias{mdpi_article}
\title{MDPI journal format.}
\usage{
mdpi_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to Multidisciplinary Digital Publishing
Institute (MDPI) journals. Adapted from
\href{http://www.mdpi.com/authors/latex}{http://www.mdpi.com/authors/latex}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "mdpi_article", package = "rticles")
}

}
