% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peerj_article.R
\name{peerj_article}
\alias{peerj_article}
\title{PeerJ journal format.}
\usage{
peerj_article(..., keep_tex = TRUE, base_format = rmarkdown::pdf_document)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{base_format}{The function to use for the base format of the article. 
By default, this is \code{rmarkdown::pdf_document}, but to use pandoc's 
cross-referencing feature, this can be set to \code{bookdown::pdf_document2}}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to The PeerJ.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "peerj_article", package = "rticles")
}

}
