% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peerj_article.R
\name{peerj_article}
\alias{peerj_article}
\title{PeerJ journal format.}
\usage{
peerj_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document()}.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\description{
Format for creating submissions to The PeerJ.
}
\details{
This was adapted from the
\href{https://www.overleaf.com/latex/templates/latex-template-for-peerj-journal-and-pre-print-submissions/ptdwfrqxqzbn}{PeerJ Overleaf Template}.
}
