% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peerj_article.R
\name{peerj_article}
\alias{peerj_article}
\title{PeerJ journal format.}
\usage{
peerj_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to The PeerJ.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "peerj_article", package = "rticles")
}

}
