% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plos_article.R
\name{plos_article}
\alias{plos_article}
\title{PLoS Computational Biology format.}
\usage{
plos_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating an article for submission to PLoS Computational Biology.
Adapted from
\href{http://journals.plos.org/plosone/s/latex}{http://journals.plos.org/plosone/s/latex}.
}

