% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnas_article.R
\name{pnas_article}
\alias{pnas_article}
\title{PNAS journal format.}
\usage{
pnas_article(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to PNAS journals.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "pnas_article", package = "rticles")
}

}
