% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjournal_article.R
\name{rjournal_article}
\alias{rjournal_article}
\title{R Journal format.}
\usage{
rjournal_article(..., keep_tex = TRUE, citation_package = "natbib")
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is now defunct in favor of the \href{https://rjournal.github.io/rjtools/}{\strong{rjtools}} package
which is now officialy recommanded by R Journal \url{https://rjournal.github.io/submissions.html}.
}
