% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjournal_article.R
\name{rjournal_article}
\alias{rjournal_article}
\title{R Journal format.}
\usage{
rjournal_article(..., citation_package = "natbib")
}
\arguments{
\item{..., citation_package}{Arguments to \code{rmarkdown::pdf_document}.}
}
\description{
Format for creating R Journal articles. Adapted from
\url{https://journal.r-project.org/submissions.html}. In order to ease the
manuscript preparation, the output is collected in a separate output directory,
which can be used for the submission. If needed file names are renamed automatically.
}
\details{
\strong{The \code{author} field in the YAML header}

\tabular{lll}{
FIELD \tab TYPE \tab DESCRIPTION\cr
\code{name} \tab \emph{required} \tab name and surname of the author\cr
\code{affiliation} \tab \emph{required} \tab name of the author's affiliation\cr
\code{address} \tab \emph{required} \tab at least one address line for the affiliation\cr
\code{url} \tab \emph{optional} \tab an additional url for the author or the main affiliation\cr
\code{orcid} \tab \emph{optional} \tab the authors ORCID if available\cr
\code{email} \tab \emph{required} \tab the author's e-mail address\cr
\code{affiliation2} \tab \emph{optional} \tab name of the author's 2nd affiliation\cr
\code{address2} \tab \emph{optional} \tab address lines belonging to the author's 2nd affiliation
}

\emph{Please note: Only one \code{url}, \code{orcid} and \code{email} can be provided per author.}

\strong{Other YAML fields}

\tabular{lll}{
FIELD \tab TYPE \tab DESCRIPTION\cr
\code{bibliography} \tab \emph{with default} \tab the BibTeX file with the reference entries\cr
\code{output_dir} \tab \emph{with default} \tab output directory where all the files are stored ready for submission
}
}
