% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsos_article.R
\name{rsos_article}
\alias{rsos_article}
\title{Royal Society Open Science journal format.}
\usage{
rsos_article(..., keep_tex = TRUE, latex_engine = "xelatex",
  pandoc_args = NULL, includes = NULL, fig_crop = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", and "xelatex".}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to Royal Society Open Science journals.
}
\author{
Thierry Onkelinx, \email{thierry.onkelinx@inbo.be}
}
