% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_article.R
\name{rss_article}
\alias{rss_article}
\title{Royal Statistical Society Journal Format}
\usage{
rss_article(..., keep_tex = TRUE, citation_package = "natbib")
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{default} if neither package is to be used,
which means citations will be processed via the command
\command{pandoc-citeproc}.}
}
\description{
Format for creating articles for Royal Statistical Society adapted from
\url{https://rss.org.uk/news-publication/publications/journals/submit-paper/}.
}
