% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rticles-package.R
\docType{package}
\name{rticles-package}
\alias{rticles}
\alias{rticles-package}
\title{rticles: Article Formats for R Markdown}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A suite of custom R Markdown formats and templates for authoring journal articles and conference submissions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/rticles}
  \item Report bugs at \url{https://github.com/rstudio/rticles/issues}
}

}
\author{
\strong{Maintainer}: Christophe Dervieux \email{cderv@posit.co} (\href{https://orcid.org/0000-0003-4474-2498}{ORCID})

Authors:
\itemize{
  \item JJ Allaire \email{jj@posit.co}
  \item Yihui Xie \email{xie@yihui.name} (\href{https://orcid.org/0000-0003-0645-5666}{ORCID})
  \item  R Foundation [copyright holder]
  \item Hadley Wickham \email{hadley@posit.co}
  \item  Journal of Statistical Software [copyright holder]
  \item Ramnath  Vaidyanathan \email{ramnath.vaidya@gmail.com} [copyright holder]
  \item  Association for Computing Machinery [copyright holder]
  \item Carl Boettiger \email{cboettig@gmail.com} [copyright holder]
  \item  Elsevier [copyright holder]
  \item Karl Broman \email{kbroman@gmail.com} [copyright holder]
  \item Kirill Mueller \email{kirill.mueller@ivt.baug.ethz.ch} [copyright holder]
  \item Bastiaan Quast \email{bquast@gmail.com} [copyright holder]
  \item Randall  Pruim \email{rpruim@gmail.com} [copyright holder]
  \item Ben Marwick \email{bmarwick@uw.edu} [copyright holder]
  \item Charlotte Wickham \email{cwickham@gmail.com} [copyright holder]
  \item Oliver Keyes \email{okeyes@wikimedia.org} [copyright holder]
  \item Miao Yu \email{yufreecas@gmail.com} [copyright holder]
  \item Daniel Emaasit \email{daniel.emaasit@gmail.com} [copyright holder]
  \item Thierry Onkelinx \email{thierry.onkelinx@inbo.be} [copyright holder]
  \item Alessandro Gasparini \email{ellessenne@gmail.com} (\href{https://orcid.org/0000-0002-8319-7624}{ORCID}) [copyright holder]
  \item Marc-Andre Desautels \email{marc-andre.desautels@cstjean.qc.ca} [copyright holder]
  \item Dominik Leutnant \email{leutnant@fh-muenster.de} (\href{https://orcid.org/0000-0003-3293-2315}{ORCID}) [copyright holder]
  \item  MDPI [copyright holder]
  \item  Taylor and Francis [copyright holder]
  \item Oğuzhan Öğreden (\href{https://orcid.org/0000-0002-9949-3348}{ORCID})
  \item Dalton Hance \email{dhance@usgs.gov}
  \item Daniel Nüst \email{daniel.nuest@uni-muenster.de} (\href{https://orcid.org/0000-0002-0024-5046}{ORCID}) [copyright holder]
  \item Petter Uvesten \email{petter.uvesten.7362@student.lu.se} [copyright holder]
  \item Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID}) [copyright holder]
  \item John Muschelli \email{muschellij2@gmail.com} (\href{https://orcid.org/0000-0001-6469-1750}{ORCID}) [copyright holder]
  \item Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID})
  \item Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID}) [copyright holder]
  \item Noam Ross \email{noam.ross@gmail.com} (\href{https://orcid.org/0000-0002-2136-0000}{ORCID}) [copyright holder]
  \item Robrecht Cannoodt \email{rcannood@gmail.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID}) (rcannood) [copyright holder]
  \item Duncan Luguern \email{duncan.luguern@gmail.com}
  \item David M. Kaplan \email{dmkaplan2000@gmail.com} (\href{https://orcid.org/0000-0001-6087-359X}{ORCID}) (dmkaplan2000) [contributor]
  \item Sebastian Kreutzer \email{sebastian.kreutzer@aber.ac.uk} (\href{https://orcid.org/0000-0002-0734-2199}{ORCID})
  \item Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID}) [contributor]
  \item Jay Hesselberth \email{jay.hesselberth@gmail.com} (\href{https://orcid.org/0000-0002-6299-179X}{ORCID}) [contributor]
  \item Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Alfredo Hernández \email{aldomann.designs@gmail.com} (\href{https://orcid.org/0000-0002-2660-4545}{ORCID}) [contributor]
  \item Stefano Coretta \email{stefano.coretta@gmail.com} (\href{https://orcid.org/0000-0001-9627-5532}{ORCID}) (stefanocoretta) [contributor]
  \item Greg Macfarlane \email{gregmacfarlane@gmail.com} (gregmacfarlane) [contributor]
  \item Matthias Templ \email{matthias.templ@gmail.com} (\href{https://orcid.org/0000-0002-8638-5276}{ORCID}) (matthias-da) [contributor]
  \item Alvaro Uzaheta \email{alvaro.uzaheta@gess.ethz.ch} (auzaheta) [contributor]
  \item JooYoung Seo \email{jseo1005@illinois.edu} (\href{https://orcid.org/0000-0002-4064-6012}{ORCID}) [contributor]
  \item Callum Arnold \email{cal.rk.arnold@gmail.com} (arnold-c) [contributor]
  \item Dmytro Perepolkin \email{dperepolkin@gmail.com} (\href{https://orcid.org/0000-0001-8558-6183}{ORCID}) (dmi3kno) [contributor]
  \item Tom Palmer \email{remlapmot@hotmail.com} (\href{https://orcid.org/0000-0003-4655-4511}{ORCID}) (remlapmot) [contributor]
}

}
\keyword{internal}
