% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/science_article.R
\name{science_article}
\alias{science_article}
\title{Science Journal Format}
\usage{
science_article(..., keep_tex = TRUE, number_sections = FALSE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{number_sections}{\code{TRUE} to number section headings}
}
\description{
Format for creating submissions to Science. Based on the Science
\href{https://www.sciencemag.org/site/feature/contribinfo/prep/TeX_help/#downloads}{class}.
}
