% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/science_article.R
\name{science_article}
\alias{science_article}
\title{Science Journal Format}
\usage{
science_article(
  ...,
  keep_tex = TRUE,
  move_figures = TRUE,
  move_tables = TRUE,
  number_sections = FALSE,
  draft = TRUE,
  pandoc_args = NULL
)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF.
Note that this argument does not control whether to keep the auxiliary
files (e.g., \file{.aux}) generated by LaTeX when compiling \file{.tex} to
\file{.pdf}. To keep these files, you may set \code{options(tinytex.clean =
FALSE)}.}

\item{move_figures}{set to \code{TRUE} to move figures to end. Default is \code{TRUE}. Only
works when \code{draft == TRUE}.}

\item{move_tables}{set to \code{TRUE} to move tables to end. Default is \code{TRUE}. Only
works when \code{draft == TRUE}.}

\item{number_sections}{\code{TRUE} to number section headings}

\item{draft}{set to \code{TRUE} for the draft version or \code{FALSE} for a
final submission version. \code{TRUE} moves the supplemental
materials to its own file.}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\description{
Format for creating submissions to Science. Based on the Science
\href{https://www.sciencemag.org/site/feature/contribinfo/prep/TeX_help/#downloads}{class}.
Note that this template only works for \code{bibtext} with \code{natbib} for citations.
}
