% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_ebook.R
\name{tufte_ebook}
\alias{tufte_ebook}
\title{Tufte e-Book format (PDF)}
\usage{
tufte_ebook(toc = TRUE, toc_depth = 3, number_sections = TRUE,
  fig_width = 4, fig_height = 2.5, highlight = "pygments",
  latex_engine = "xelatex", ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_depth}{Depth of headers to include in table of contents}

\item{number_sections}{\code{TRUE} to number section headings}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", and "xelatex".  Note that lualatex may have
problems with text spacing, and that pdflatex may have memory issues when
using tikzDevice.}

\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}
}
\description{
Format for creating an e-book based on the style of Edward R. Tufte and
Richard Feynman.
}

